# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TokenTimer Core API Contract
  version: 0.1.0
  description: |
    TokenTimer Core API for authentication, workspaces, token lifecycle management,
    alerts, audit, and integration import/scan workflows.
    This contract is also used as a compatibility baseline across core, cloud, and enterprise variants.
servers:
  - url: http://localhost:4000
    description: Local development backend
...
security:
  - cookieAuth: []
paths:
...
  /api/test-whatsapp:
    post:
      tags: [Alerts]
      summary: Send workspace WhatsApp test message
      operationId: sendWorkspaceWhatsAppTest
      description: Sends a workspace-scoped WhatsApp test message with cooldown and validation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workspace_id:
                  type: string
                  format: uuid
                phone_e164:
                  type: string
      responses:
        "200":
          description: WhatsApp test result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: Cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /webhooks/twilio/whatsapp/status:
    post:
      tags: [Alerts]
      summary: Twilio WhatsApp delivery status webhook
      operationId: twilioWhatsAppStatusWebhook
      description: Receives Twilio status callbacks and updates WhatsApp delivery metadata.
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook processed
...
  /api/v1/workspaces/{id}/tokens/reassign-contact-group:
    post:
      tags: [Workspaces]
      summary: Reassign tokens between contact groups
      operationId: reassignTokenContactGroup
      description: Moves token contact_group_id values from one group to another in the workspace.
      parameters:
        - $ref: "#/components/parameters/workspaceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_group_id, to_group_id]
              properties:
                from_group_id:
                  type: string
                to_group_id:
                  type: string
      responses:
        "200":
          description: Reassignment result
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session cookie authentication used by authenticated API routes.
  parameters:
    workspaceIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          oneOf:
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true
            - type: string
      example:
        error: Invalid request payload
        code: VALIDATION_ERROR
        details:
          - email is required
...